+++ src/Executables/nesWorkerStarter.cpp
@@ -66,14 +66,14 @@ int main(int argc, char** argv) {
 
     std::string sourceType = "";
     std::string sourceConfig = "";
-    size_t sourceFrequency = 0;
-    size_t numberOfBuffersToProduce = 1;
+    uint64_t sourceFrequency = 0;
+    uint64_t numberOfBuffersToProduce = 1;
     std::string physicalStreamName;
     std::string logicalStreamName;
     std::string parentId = "-1";
     std::string endlessRepeat = "";
     std::string skipHeader = "false";
-    size_t numberOfTuplesToProducePerBuffer = 0;
+    uint64_t numberOfTuplesToProducePerBuffer = 0;
 
     desc.add_options()("coordinatorPort", po::value<string>(&coordinatorPort)->default_value(coordinatorPort),
                        "Set NES rpc server port (default: 0).")("rpcPort", po::value<string>(&rpcPort)->default_value(rpcPort),
@@ -84,14 +84,14 @@ int main(int argc, char** argv) {
                                                    "Set the type of the Source either CSVSource or DefaultSource")(
         "sourceConfig", po::value<string>(&sourceConfig)->default_value(sourceConfig),
         "Set the config for the source e.g. the file name")(
-        "sourceFrequency", po::value<size_t>(&sourceFrequency)->default_value(sourceFrequency), "Set the sampling frequency")(
+        "sourceFrequency", po::value<uint64_t>(&sourceFrequency)->default_value(sourceFrequency), "Set the sampling frequency")(
         "endlessRepeat", po::value<string>(&endlessRepeat)->default_value("off"), "Looping endless over the file")(
         "skipHeader", po::value<string>(&skipHeader)->default_value("false"), "Skip first line of the file (default=false)")(
         "physicalStreamName", po::value<string>(&physicalStreamName)->default_value(physicalStreamName),
         "Set the physical name of the stream")(
-        "numberOfBuffersToProduce", po::value<size_t>(&numberOfBuffersToProduce)->default_value(numberOfBuffersToProduce),
+        "numberOfBuffersToProduce", po::value<uint64_t>(&numberOfBuffersToProduce)->default_value(numberOfBuffersToProduce),
         "Set the number of buffers to produce")("numberOfTuplesToProducePerBuffer",
-                                                po::value<size_t>(&numberOfTuplesToProducePerBuffer)->default_value(0),
+                                                po::value<uint64_t>(&numberOfTuplesToProducePerBuffer)->default_value(0),
                                                 "Set the number of buffers to produce")(
         "logicalStreamName", po::value<string>(&logicalStreamName)->default_value(logicalStreamName),
         "Set the logical stream name where this stream is added to")(
@@ -129,8 +129,8 @@ int main(int argc, char** argv) {
 
     // TODO remote calls to cout
 
-    size_t localPort = std::stol(rpcPort);
-    size_t zmqDataPort = std::stol(dataPort);
+    uint64_t localPort = std::stol(rpcPort);
+    uint64_t zmqDataPort = std::stol(dataPort);
     cout << "port=" << localPort << "localport=" << std::to_string(localPort) << " pid=" << getpid() << endl;
     NesWorkerPtr wrk =
         std::make_shared<NesWorker>(coordinatorIp, coordinatorPort, localWorkerIp, localPort, zmqDataPort, numberOfSlots,
